Bugfix metatable applied to `_ENV`
Accessing a dynamic value in the `env` table causes a function call to generate the value. The `_ENV` metatable invoked `env_has(key)`, which performs a lookup of `env[key]` thereby potentially invoking a dynamic value's function call, and if a value is found, re-accesses `env[key]` to produce the value. When accessing a dynamic value in this way, the defined function would be invoked twice: once to check if a value exists, and again to actually return it.